api/requirements.txt
flask~=3.0.3
protobuf~=5.28.2
google-api-python-client~=2.149.0